Think it all works?

+++ client/src/com/github/blahblahbloopster/crypto/CommunicationSystem.kt
@@ -8,8 +8,8 @@ interface CommunicationSystem {
     val id: Int
     /** The maximum number of bytes that can be sent at once. */
     val MAX_LENGTH: Int
-    /** The time in milliseconds to wait between transmissions. */
-    val RATE: Long
+    /** The time in 1/60s to wait between transmissions. */
+    val RATE: Float
 
     /** Initializes the system. */
     fun init() {}
